celery_app/requirements.txt
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
psycopg2-binary==2.9.7
requests==2.31.0